Add building gtest for OpenCL in GPU build

+++ tests/cpp-runtime/opencl/opencl_texture_pool_test.cc
@@ -31,13 +31,13 @@ using namespace tvm::runtime::cl;
 // get and check internal state of class Pool
 class PoolWrapper : public Pool2D {
  public:
-  inline size_t FreeListSize() const { return free_list_.size(); }
-  inline size_t AllocatedListSize() const { return allocated_.size(); }
-  inline std::pair<size_t, size_t> FreeListItemSize(size_t idx) const {
-    return std::make_pair(free_list_[idx].x, free_list_[idx].y);
+  inline int FreeListSize() const { return static_cast<int>(free_list_.size()); }
+  inline int AllocatedListSize() const { return static_cast<int>(allocated_.size()); }
+  inline std::pair<int, int> FreeListItemSize(size_t idx) const {
+    return std::make_pair(static_cast<int>(free_list_[idx].x), static_cast<int>(free_list_[idx].y));
   }
-  inline std::pair<size_t, size_t> AllocatedListItemSize(size_t idx) const {
-    return std::make_pair(allocated_[idx].x, allocated_[idx].y);
+  inline std::pair<int, int> AllocatedListItemSize(size_t idx) const {
+    return std::make_pair(static_cast<int>(allocated_[idx].x), static_cast<int>(allocated_[idx].y));
   }
 };
